11031 fix tests

+++ src/api2/routes/shamirRecovery/openShamirShares.ts
@@ -73,8 +73,8 @@ export const openShamirShares = async (req: Request, res: Response): Promise<voi
     await db.query(
       `UPDATE shamir_recovery_requests
           SET approved_by = CASE
-            WHEN approved_by IS NULL THEN ARRAY[$1]
-            WHEN NOT ($1 = ANY(approved_by)) THEN array_append(approved_by, $1)
+            WHEN approved_by IS NULL THEN ARRAY[$1::integer]
+            WHEN NOT ($1::integer = ANY(approved_by)) THEN array_append(approved_by, $1::integer)
             ELSE approved_by
           END
         WHERE id = $2
